requirements.txt
#django
Django==3.2
djangorestframework==3.13.1
Pillow==9.2.0

